backend/requirements.txt
# FastAPI + ASGI
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
pydantic[email]==2.7.4
pydantic-settings==2.3.1

# HTTP Client
httpx==0.27.0

# Workers
celery==5.4.0
redis==5.0.7

# Utilities
python-dotenv==1.0.1

# Dev / Test
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
